+++ src/main/java/eu/midnightdust/blur/animations/impl/FadeAnimationHandler.java
@@ -2,7 +2,6 @@ package eu.midnightdust.blur.animations.impl;
 
 import eu.midnightdust.blur.animations.AbstractAnimationHandler;
 import eu.midnightdust.blur.animations.AnimationState;
-import eu.midnightdust.blur.animations.FadeAnimationState;
 import eu.midnightdust.blur.animations.IAnimationHandler;
 import eu.midnightdust.blur.config.BlurConfig;
