+++ src/main/java/eu/midnightdust/blur/animations/impl/GradientAnimationHandler.java
@@ -2,7 +2,6 @@ package eu.midnightdust.blur.animations.impl;
 
 import eu.midnightdust.blur.animations.AbstractAnimationHandler;
 import eu.midnightdust.blur.animations.AnimationState;
-import eu.midnightdust.blur.animations.GradientAnimationState;
 import eu.midnightdust.blur.animations.IAnimationHandler;
 import eu.midnightdust.blur.config.BlurConfig;
 import eu.midnightdust.lib.util.MidnightColorUtil;
