+++ src/main/java/eu/midnightdust/blur/mixin/MixinTitleScreen.java
@@ -1,7 +1,7 @@
 package eu.midnightdust.blur.mixin;
 
 import eu.midnightdust.blur.Blur;
-import eu.midnightdust.blur.animations.FadeAnimationState;
+import eu.midnightdust.blur.animations.impl.FadeAnimationState;
 import eu.midnightdust.blur.config.BlurConfig;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.Screen;
